refactor(server): trim INSTRUCTIONS + extract envelope/render helpers + sync README tool count (closes #101, closes #102)
- INSTRUCTIONS: 178 → 41 lines. Remove McKinsey-specific layout rules,
  data density guidelines, color palette prose, and slide-spec / chart /
  icon / callout worked examples. Keep only parameter conventions, workflow
  guidance, theme enumeration, response-shape docs, auto-render env vars.
- Extract structured-response helpers (`_success`, `_error`, `_err`,
  `_success_with_render`) into `pptx_mcp_server._envelope`.
- Extract auto-render gate (`_auto_render_enabled`, `_auto_render_timeout`,
  plus timeout-enforced runner) into `pptx_mcp_server._render`. server.py
  keeps a thin `_auto_render` shim that injects its own `render_slide`
  binding so existing `monkeypatch.setattr(server, "render_slide", …)` tests
  stay green.
- Helper modules are module-private (underscore-prefixed) — not public API.
- server.py: 1371 → 1111 LOC.
- README.md: tool count 25 → 37; add rows for `pptx_move_slide`,
  `pptx_add_auto_fit_textbox`, `pptx_add_flex_container`,
  `pptx_add_responsive_card_row`, `pptx_build_deck`, `pptx_build_slide`,
  `pptx_add_connector`, `pptx_add_callout`, `pptx_add_chart`,
  `pptx_add_icon`, `pptx_list_icons`, `pptx_check_layout`. README table rows
  now match `@mcp.tool()` registrations 1:1.
- composites.py audited — already cleanly separated (pure `_*` helpers take
  `prs`/`slide`; public wrappers do the open/save). No change required.
- No behaviour change; 638 passed / 1 skipped (same as main).
